+++ Crypto/Flows/Screens/SettingsScreen/SettingsView.swift
@@ -1,0 +1,22 @@
+//
+//  SettingsView.swift
+//  Crypto
+//
+//  Created by Developer on 01.12.2022.
+//
+
+import SwiftUI
+
+struct SettingsView: View {
+    var body: some View {
+        VStack {
+            
+        }
+    }
+}
+
+struct SettingsView_Previews: PreviewProvider {
+    static var previews: some View {
+        SettingsView()
+    }
+}
